Add assets bind thru cookie-cutter template
